feat(ide): dual-boot the IDE from a template or a GitHub repo
Extract the mode-agnostic IdeShell from the /ide route, add a landing
screen with a framework grid and a GitHub URL box, and add the
/ide/github/[owner]/[repo]/tree/[ref]/[...dir] route. Framework
switching moves from the old side panel into a header project
switcher.
